Module minewars::camera

General camera stuff

Because we want to make the ALPHA playtesting prototype with 2D graphics, and then transition to 3D graphics in future versions of the game, any behavior related to the 2D game camera is in gfx2d::camera. This file contains stuff that is agnostic between 2D/3D.

Structs

CameraControlSet

CameraJumpTo

Event to cause a (smooth) jump to a given coordinate position

GameCamera

Marker for the main game camera (that renders the gameplay map view)

GridCursor

GridCursorSet

MwCameraPlugin

Enums

ScreenShakeEvent

Event to cause a screen shake