Module minewars::prelude

Convenience, to be imported in every file in the crate

Modules

helpers

A module which provides helper functions.

num

A collection of numeric types and traits for Rust.

num_traits

Numeric traits for generic mathematics

quinn

QUIC transport protocol support for Tokio

rand

Utilities for random number generation

rand_pcg

The PCG random number generators.

rustls

Rustls - a modern TLS library

shape

Generation for some primitive shape meshes.

shapes

Collection of common shapes that can be drawn.

system_adapter

A collection of common adapters for piping the result of a system.

tess

Tessellation of 2D fill and stroke operations.

tokio

A runtime for writing reliable network applications without compromising speed.

Macros

anyhow

Construct an ad-hoc error from a string or existing non-anyhow error value.

bail

Return early with an error.

debug

Constructs an event at the debug level.

debug_span

Constructs a span at the debug level.

ensure

error

Constructs an event at the error level.

error_span

Constructs a span at the error level.

info

Constructs an event at the info level.

info_span

Constructs a span at the info level.

trace

Constructs an event at the trace level.

trace_span

Constructs a span at the trace level.

warn

Constructs an event at the warn level.

warn_span

Constructs a span at the warn level.

Structs

AabbGizmo

Add this Component to an entity to draw its Aabb component.

AabbGizmoConfig

Configuration for drawing the Aabb component on entities.

AddChild

Command that adds a child to an entity.

Added

A filter on a component that only retains results added after the system last ran.

AllSettings

AmbientLight

An ambient light, which lights the entire scene equally.

AncestorIter

An Iterator of Entitys over the ancestors of an Entity.

AnimatedTile

A component that is attached to a Tile entity that tells the GPU how to animate the tile. Currently all frames must be aligned in your tilemap.

AnimationClip

A list of VariableCurve, and the EntityPath to which they apply.

AnimationPlayer

Animation controls

AnimationPlugin

Adds animation support to an app

AnyError

The Error type, a wrapper around a dynamic error type.

AnyOf

The AnyOf query parameter fetches entities with any of the component types included in T.

App

A container of app logic and data.

AppStatesPlugin

AppTypeRegistry

A Resource storing TypeRegistry for type registrations relevant to a whole app.

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

ArrayTextureLoader

A bevy world resource that allows you to add atlas textures for loading into the array texture cache.

AssetPlugin

Adds support for Assets to an App.

AssetServer

Loads assets from the filesystem in the background.

Assets

Stores Assets of a given type and tracks changes to them.

AtlasImageBundle

A UI node that is a texture atlas sprite

AudioSink

Used to control audio during playback.

AudioSource

A source of audio data

AudioSourceBundle

Bundle for playing a sound.

Axis

Stores the position data of the input devices of type T.

BSpline

A spline interpolated continuously across the nearest four control points. The curve does not pass through any of the control points.

BVec2

A 2-dimensional bool vector mask.

BVec3

A 3-dimensional bool vector mask.

BVec4

A 4-dimensional bool vector mask.

BackgroundColor

The background color of the node

Bezier

A spline composed of a series of cubic Bezier curves.

BorderColor

The border color of the UI node.

Button

Marker struct for buttons

ButtonBundle

A UI node that is a button

CalculatedClip

The calculated clip of the node

Camera

The defining Component for camera entities, storing information about how and what to render through this camera.

Camera2d

Camera2dBundle

Camera3d

Configuration for the “main 3d render graph”.

Camera3dBundle

CardinalSpline

A spline interpolated continuously across the nearest four control points, with the position of the curve specified at every control point and the tangents computed automatically.

ChainOptionalSystem

Similar to Bevy’s PipedSystem, but the second system is optional

ChainResultSystem

Similar to Bevy’s PipedSystem, but with diverging paths for Ok/Err Results

Changed

A filter on a component that only retains results added or mutably dereferenced after the system last ran.

ChildBuilder

Struct for building children entities and adding them to a parent entity.

Children

Contains references to the child entities of this entity.

ClearChildren

Command that clears all children from an entity and removes Parent component from those children.

ClearColor

A Resource that stores the color that is used to clear the screen between frames.

ColorMaterial

A 2d material that renders 2d meshes with a texture tinted by a uniform color

Commands

A Command queue to perform impactful changes to the World.

ComputedVisibility

Algorithmically-computed indication of whether an entity is visible and should be extracted for rendering

CubicSegment

A segment of a cubic curve, used to hold precomputed coefficients for fast interpolation.

CursorEntered

An event that is sent whenever the user’s cursor enters a window.

CursorLeft

An event that is sent whenever the user’s cursor leaves a window.

CursorMoved

An event reporting that the mouse cursor has moved inside a window.

DebugName

Convenient query for giving a human friendly name to an entity.

DefaultPlugins

This plugin group will add all the default plugins for a Bevy application:

Deferred

A SystemParam that stores a buffer which gets applied to the World during apply_deferred. This is used internally by Commands to defer World mutations.

DescendantIter

An Iterator of Entitys over the descendants of an Entity.

DespawnChildrenRecursive

Despawns the given entity’s children recursively

DespawnRecursive

Despawns the given entity and all its children recursively

DirectionalLight

A Directional light.

DirectionalLightBundle

A component bundle for DirectionalLight entities.

Duration

A Duration type to represent a span of time, typically used for system timeouts.

DynamicAssetCollections

Resource keeping track of dynamic asset collection files for different loading states

DynamicAssets

Resource to dynamically resolve keys to assets.

DynamicScene

A collection of serializable resources and dynamic entities.

DynamicSceneBuilder

A DynamicScene builder, used to build a scene from a World by extracting some entities and resources.

DynamicSceneBundle

A component bundle for a DynamicScene root.

Entity

Lightweight identifier of an entity.

EntityPath

Path to an entity, with Names. Each entity in a path must have a name.

EntityRef

A read-only reference to a particular Entity and all of its components

EnvironmentMapLight

Environment map based ambient lighting representing light from distant scenery.

EventReader

Reads events of type T in order and tracks which events have already been read.

EventWriter

Sends events of type T.

Events

An event collection that represents the events that occurred within the last two Events::update calls. Events can be written to using an EventWriter and are typically cheaply read using an EventReader.

ExtractSchedule

Schedule which extract data from the main world and inserts it into the render world.

Fill

Defines the fill options for the lyon tessellator and color of the generated vertices.

FillOptions

Parameters for the fill tessellator.

First

Runs first in the schedule. This is run by the Main schedule.

FixedTime

The amount of time that must pass before the fixed timestep schedule is run again.

FixedUpdate

The schedule that contains systems which only run after a fixed period of time has elapsed.

FogSettings

Configures the “classic” computer graphics distance fog effect, in which objects appear progressively more covered in atmospheric haze the further away they are from the camera. Affects meshes rendered via the PBR StandardMaterial.

Font

FrameCountPlugin

Adds frame counting functionality to Apps.

Gamepad

A gamepad with an associated ID.

GamepadAxis

An axis of a Gamepad.

GamepadButton

A button of a Gamepad.

Gamepads

A collection of connected Gamepads.

GeometryBuilder

Allows the creation of shapes using geometries added to a path builder.

GilrsPlugin

GizmoConfig

A Resource that stores configuration for gizmos.

Gizmos

A SystemParam for drawing gizmos.

GlobalTransform

Describe the position of an entity relative to the reference frame.

GlobalVolume

Use this Resource to control the global volume of all audio with a Volume::Relative volume.

GridPlacement

Represents the position of a grid item in a single axis.

GridTrack

A GridTrack is a Row or Column of a CSS Grid. This struct specifies what size the track should be. See below for the different “track sizing functions” you can specify.

Handle

A handle into a specific Asset of type T.

HandleUntyped

A non-generic version of Handle.

Hermite

A spline interpolated continuously between the nearest two control points, with the position and velocity of the curve specified at both control points. This curve passes through all control points, with the specified velocity which includes direction and parametric speed.

HiddenProgress

“Hidden” progress reported by a system.

HierarchyPlugin

The base plugin for handling Parent and Children components

IVec2

A 2-dimensional vector.

IVec3

A 3-dimensional vector.

IVec4

A 4-dimensional vector.

Image

ImageBundle

A UI node that is an image

ImagePlugin

Adds the Image as an asset and makes sure that they are extracted and prepared for the GPU.

In

Wrapper type to mark a SystemParam as an input.

InGameSet

InStateSet

Input

A “press-able” input of type T.

InsertChildren

Command that inserts a child at a given index of a parent’s children, shifting following children back.

Label

Marker struct for labels

Last

Runs last in the schedule. This is run by the Main schedule.

LoadingState

A Bevy plugin to configure automatic asset loading

Local

A system local SystemParam.

Main

The schedule that contains the app logic that is evaluated each tick of App::update().

Mat2

A 2x2 column major matrix.

Mat3

A 3x3 column major matrix.

Mat4

A 4x4 column major matrix.

MaterialMeshBundle

A component bundle for entities with a Mesh and a Material.

MaterialPlugin

Adds the necessary ECS resources and render logic to enable rendering entities using the given Material asset type.

MaterialTilemapBundle

The default tilemap bundle. All of the components within are required.

MaterialTilemapPlugin

Mesh

A 3D object made out of vertices representing triangles, lines, or points, with “attribute” values for each vertex.

MinimalPlugins

This plugin group will add the minimal plugins for a Bevy application:

MorphWeights

Controls the morph targets for all child Handle<Mesh> entities. In most cases, MorphWeights should be considered the “source of truth” when writing morph targets for meshes. However you can choose to write child MeshMorphWeights if your situation requires more granularity. Just note that if you set MorphWeights, it will overwrite child MeshMorphWeights values.

Mut

Unique mutable borrow of an entity’s component

Name

Component used to identify an entity. Stores a hash for faster comparisons.

NeedsSettingsSet

NextState

The next state of State<S>.

Node

Describes the size of a UI node

NodeBundle

The basic UI node

NonSend

Shared borrow of a non-Send resource.

NonSendMut

Unique borrow of a non-Send resource.

OnClick

Add this component to an interactable UI node, to specify things to run automatically on click.

OnEnter

The label of a Schedule that runs whenever State<S> enters this state.

OnExit

The label of a Schedule that runs whenever State<S> exits this state.

OnTransition

The label of a Schedule that only runs whenever State<S> exits the from state, AND enters the to state.

Or

A filter that tests if any of the given filters apply.

OrthographicProjection

Project a 3D space onto a 2D surface using parallel lines, i.e., unlike PerspectiveProjection, the size of objects remains the same regardless of their distance to the camera.

Overflow

Whether to show or hide overflowing items

ParallelCommands

An alternative to Commands that can be used in parallel contexts, such as those in Query::par_iter

ParamSet

A collection of potentially conflicting SystemParams allowed by disjoint access.

Parent

Holds a reference to the parent entity of this entity. This component should only be present on entities that actually have a parent entity.

PathBuf

An owned, mutable path (akin to String).

PathBuilder

A SVG-like path builder.

PerspectiveProjection

A 3D camera projection in which distant objects appear smaller than close objects.

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

PlaybackSettings

Initial settings to be used when audio starts playing. If you would like to control the audio while it is playing, query for the AudioSink or SpatialAudioSink components. Changes to this component will not be applied to already-playing audio.

PointLight

A light that emits light in all directions from a central point.

PointLightBundle

A component bundle for PointLight entities.

PostStartup

The schedule that runs once after Startup. This is run by the Main schedule.

PostUpdate

The schedule that contains logic that must run after Update. For example, synchronizing “local transforms” in a hierarchy to “global” absolute transforms. This enables the PostUpdate transform-sync system to react to “local transform” changes in Update without the Update systems needing to know about (or add scheduler dependencies for) the “global transform sync system”.

PreStartup

The schedule that runs before Startup. This is run by the Main schedule.

PreUpdate

The schedule that contains logic that must run before Update. For example, a system that reads raw keyboard input OS events into an Events resource. This enables systems in Update to consume the events from the Events resource without actually knowing about (or taking a direct scheduler dependency on) the “os-level keyboard event sytsem”.

Progress

Progress reported by a system

ProgressCounter

Resource for tracking overall progress

ProgressPlugin

Add this plugin to your app, to use this crate for the specified state.

PushChildren

Command that pushes children to the end of the entity’s Children.

Quat

A quaternion representing an orientation.

Query

System parameter that provides selective access to the Component data stored in a World.

QueryState

Provides scoped access to a World state according to a given WorldQuery and query filter.

Ray

A ray is an infinite line starting at origin, going in direction.

ReceivedCharacter

An event that is sent whenever a window receives a character from the OS or underlying system.

Rect

A rectangle defined by two opposite corners.

Ref

Shared borrow of an entity’s component with access to change detection. Similar to Mut but is immutable and so doesn’t require unique access.

ReflectComponent

A struct used to operate on reflected Component of a type.

ReflectDefault

A struct used to provide the default value of a type.

ReflectDeserialize

A struct used to deserialize reflected instances of a type.

ReflectFromReflect

Type data that represents the FromReflect trait and allows it to be used dynamically.

ReflectResource

A struct used to operate on reflected Resource of a type.

ReflectSerialize

A struct used to serialize reflected instances of a type.

RegisterStandardDynamicAsset

Command to register a standard dynamic asset under the given key

RegularPolygon

RemoveChildren

Command that removes children from an entity, and removes these children’s parent.

RemoveParent

Command that removes the parent of an entity, and removes that entity from the parent’s Children.

RemovedComponents

A SystemParam that grants access to the entities that had their T Component removed.

RepeatedGridTrack

Represents a possibly repeated GridTrack.

ReplaceChildren

Command that clear all children from an entity, replacing them with the given children.

Res

Shared borrow of a Resource.

ResMut

Unique mutable borrow of a Resource.

RumbleSystem

Updates the running gamepad rumble effects.

RxBroadcast

Receiving-half of the broadcast channel.

RxMpsc

Receives values from the associated Sender.

RxMpscU

Receive values from the associated UnboundedSender.

RxOneshot

Receives a value from the associated Sender.

ScanCode

The scan code of a KeyboardInput.

Scene

To spawn a scene, you can use either:

SceneBundle

A component bundle for a Scene root.

SceneSpawner

Schedule

A collection of systems, and the metadata and executor needed to run them in a certain order under certain conditions.

Schedules

Resource that stores Schedules mapped to ScheduleLabels.

ScreenSpaceAmbientOcclusionPlugin

Plugin for screen space ambient occlusion.

Shader

A shader, as defined by its ShaderSource and ShaderStage This is an “unprocessed” shader. It can contain preprocessor directives.

ShapeBundle

A Bevy Bundle to represent a shape.

ShapePath

A builder for Paths based on shapes implementing Geometry.

ShapePlugin

A plugin that provides resources and a system to draw shapes in Bevy with less boilerplate.

SpatialAudioSink

Used to control spatial audio during playback.

SpatialAudioSourceBundle

Bundle for playing a sound with a 3D position.

SpatialBundle

A Bundle with the following Components:

SpatialSettings

Settings for playing spatial audio.

SpotLight

A light that emits light in a given direction from a central point. Behaves like a point light in a perfectly absorbent housing that shines light only in a given direction. The direction is taken from the transform, and can be specified with Transform::looking_at.

SpotLightBundle

A component bundle for spot light entities

Sprite

SpriteBundle

SpriteSheetBundle

A Bundle of components for drawing a single sprite from a sprite sheet (also referred to as a TextureAtlas)

StandardDynamicAssetCollection

The asset defining a mapping from asset keys to dynamic assets

StandardMaterial

A material with “standard” properties used in PBR lighting Standard property values with pictures here https://google.github.io/filament/Material%20Properties.pdf.

StandardTilemapMaterial

Startup

The schedule that runs once when the app starts. This is run by the Main schedule.

State

A finite-state machine whose transitions have associated schedules ([OnEnter(state)] and [OnExit(state)]).

StateDespawnMarker

Everything that must be despawned when transitioning the main app state

StateTransition

Runs state transitions. This is run by the Main schedule.

StdRng

The standard RNG. The PRNG algorithm in StdRng is chosen to be efficient on the current platform, to be statistically strong and unpredictable (meaning a cryptographically secure PRNG).

Stroke

Defines the stroke options for the lyon tessellator and color of the generated vertices.

StrokeOptions

Parameters for the tessellator.

Style

Describes the style of a UI container node

TaskPoolOptions

Helper for configuring and creating the default task pools. For end-users who want full control, set up TaskPoolPlugin

TaskPoolPlugin

Setup of default task pools: AsyncComputeTaskPool, ComputeTaskPool, IoTaskPool.

Text

Text2dBundle

The bundle of components needed to draw text in a 2D scene via a 2D Camera2dBundle. Example usage.

TextBundle

A UI node that is text

TextSection

TextStyle

TextureAtlas

An atlas containing multiple textures (like a spritesheet or a tilemap). Example usage animating sprite. Example usage loading sprite sheet.

TextureAtlasBuilder

A builder which is used to create a texture atlas from many individual sprites.

TextureAtlasSprite

ThreadRng

A reference to the thread-local generator

TileBundle

This an optional tile bundle with default components.

TileColor

A custom color for the tile.

TileFlip

Flips the tiles texture along the X, Y or diagonal axes

TilePos

A tile position in the tilemap grid.

TilePosOld

TileStorage

Used to store tile entities for fast look up. Tile entities are stored in a grid. The grid is always filled with None.

TileTextureIndex

A texture index into the atlas or texture array for a single tile. Indices in an atlas are horizontal based.

TileVisible

Hides or shows a tile based on the boolean. Default: True

TilemapArrayTexture

TilemapGridSize

Size of the tiles on the grid in pixels. This can be used to overlay tiles on top of each other. Ex. A 16x16 pixel tile can be overlapped by 8 pixels by using a grid size of 16x8.

TilemapId

A component which stores a reference to the tilemap entity.

TilemapPlugin

A bevy tilemap plugin. This must be included in order for everything to be rendered. But is not necessary if you are running without a renderer.

TilemapRenderSettings

Custom parameters for the render pipeline.

TilemapSize

Size of the tilemap in tiles.

TilemapSpacing

Spacing between tiles in pixels inside of the texture atlas. Defaults to 0.0

TilemapTextureSize

Size of the atlas texture in pixels.

TilemapTileSize

Size of the tiles in pixels

Time

A clock that tracks how much it has advanced (and how much real time has elapsed) since its previous update and since its creation.

Timer

Tracks elapsed time. Enters the finished state once duration is reached.

TouchInput

A touch input event.

Touches

A collection of Touches.

Transform

Describe the position of an entity. If the entity has a parent, the position is relative to its parent position.

TransformBundle

A Bundle of the Transform and GlobalTransform Components, which describe the position of an entity.

TransformPlugin

The base plugin for handling Transform components

TxBroadcast

Sending-half of the broadcast channel.

TxMpsc

Sends values to the associated Receiver.

TxMpscU

Send values to the associated UnboundedReceiver.

TxOneshot

Sends a value to the associated Receiver.

TypeRegistrationPlugin

Registration of default types to the TypeRegistry resource.

UVec2

A 2-dimensional vector.

UVec3

A 3-dimensional vector.

UVec4

A 4-dimensional vector.

UiCameraConfig

Configuration for cameras related to UI.

UiDisabled

For disabling UI elements; our systems for handling Interactions should look for this

UiExtrasPlugin

UiImage

The 2D texture displayed for this UI node

UiRect

A type which is commonly used to define margins, paddings and borders.

UiScale

The current scale of the UI.

UiTextureAtlasImage

The atlas sprite to be used in a UI Texture Atlas Node

Update

The schedule that contains app logic. This is run by the Main schedule.

ValidParentCheckPlugin

Print a warning for each Entity with a T component whose parent doesn’t have a T component.

VariableCurve

Describes how an attribute of a Transform or MorphWeights should be animated.

Vec2

A 2-dimensional vector.

Vec3

A 3-dimensional vector.

Vec4

A 4-dimensional vector.

VisibilityBundle

A Bundle of the Visibility and ComputedVisibility Components, which describe the visibility of an entity.

Window

The defining Component for window entities, storing information about how it should appear and behave.

WindowMoved

An event that is sent when a window is repositioned in physical pixels.

WindowPlugin

A Plugin that defines an interface for windowing support in Bevy.

WindowResizeConstraints

The size limits on a Window.

With

Filter that selects entities with a component T.

Without

Filter that selects entities without a component T.

World

Stores and exposes operations on entities, components, resources, and their associated metadata.

WorldChildBuilder

Struct for adding children to an entity directly through the World for use in exclusive systems.

WorldCursor

WorldCursorCamera

WorldCursorPlugin

WorldCursorSet

Enums

AlignContent

Defines how each line is aligned within the flexbox.

AlignItems

How items are aligned according to the cross axis

AlignSelf

How this item is aligned according to the cross axis. Overrides AlignItems.

AlphaMode

Sets how a material’s base color alpha channel is used for transparency.

AppState

State type: Which “screen” is the app in?

AssetEvent

Events that involve assets of type T.

Color

CursorIcon

The icon to display for a Window’s Cursor.

Direction

Defines the text direction

DynamicAssetType

Different typed that can generate the asset field value of a dynamic asset

EulerRot

Euler rotation sequences.

FileDragAndDrop

Events related to files being dragged and dropped on a window.

FillRule

The fill rule defines how to determine what is inside and what is outside of the shape.

FlexDirection

Defines how flexbox items are ordered within a flexbox

FlexWrap

Defines if flexbox items appear on a single line or on multiple lines

FogFalloff

Allows switching between different fog falloff modes, and configuring their parameters.

GameMode

GamepadAxisType

A type of a GamepadAxis.

GamepadButtonType

A type of a GamepadButton.

GridAutoFlow

Controls whether grid items are placed row-wise or column-wise. And whether the sparse or dense packing algorithm is used.

GridTrackRepetition

How many times to repeat a repeated grid track

HexCoordSystem

Different hex grid coordinate systems. You can find out more at this link: https://www.redblobgames.com/grids/hexagons/

Ime

A Input Method Editor event.

Interaction

Describes what type of input interaction has occurred for a UI node.

IpAddr

An IP address, either IPv4 or IPv6.

IsoCoordSystem

Different isometric coordinate systems.

JustifyContent

Defines how items are aligned according to the main axis

JustifyItems

How items are aligned according to the cross axis

JustifySelf

How this item is aligned according to the cross axis. Overrides AlignItems.

KeyCode

The key code of a KeyboardInput.

Keyframes

List of keyframes for one of the attribute of a Transform.

LineCap

Line cap as defined by the SVG specification.

LineJoin

Line join as defined by the SVG specification.

MaxTrackSizingFunction

MinTrackSizingFunction

MonitorSelection

References a screen monitor.

MouseButton

A button on a mouse device.

Msaa

Configuration resource for Multi-Sample Anti-Aliasing.

Orientation

Vertical or Horizontal.

OverflowAxis

Whether to show or hide overflowing items

ParallaxMappingMethod

The parallax mapping method to use to compute depth based on the material’s depth_map.

PositionType

The strategy used to position this node

Projection

A configurable CameraProjection that can select its projection type at runtime.

RectangleOrigin

Defines where the origin, or pivot of the Rectangle should be positioned.

RegularPolygonFeature

The regular polygon feature used to determine the dimensions of the polygon.

SceneFilter

A filter used to control which types can be added to a DynamicScene.

SessionKind

State type: What drives the game? Where do data and events come from?

SocketAddr

An internet socket address, either IPv4 or IPv6.

StandardDynamicAsset

These asset variants can be loaded from configuration files. They will then replace a dynamic asset based on their keys.

TextAlignment

Describes horizontal alignment preference for positioning & bounds.

TextError

TilemapTexture

TilemapType

The type of tile to be rendered, currently we support: Square, Hex, and Isometric.

TimerMode

Specifies Timer behavior.

Val

Represents the possible value types for layout properties.

ValArithmeticError

Visibility

User indication of whether an entity is visible. Propagates down the entity hierarchy.

WindowPosition

Defines where a Window should be placed on the screen.

ZIndex

Indicates that this Node entity’s front-to-back ordering is not controlled solely by its location in the UI hierarchy. A node with a higher z-index will appear on top of other nodes with a lower z-index.

Constants

PROPRIETARY

Traits

AddAsset

App extension methods for adding new asset types.

AssetCollection

Trait to mark a struct as a collection of assets

AssetCollectionApp

Extension trait for App enabling initialisation of asset collections

AssetCollectionWorld

Extension trait for World enabling initialisation of asset collections

AudioSinkPlayback

Common interactions with an audio sink.

BuildChildren

Trait for removing, adding and replacing children and parents of an entity.

BuildChildrenTransformExt

Collection of methods similar to [BuildChildren], but preserving each entity’s GlobalTransform.

BuildWorldChildren

Trait that defines adding, changing and children and parents of an entity directly through the World.

Bundle

The Bundle trait enables insertion and removal of Components from an entity.

CliCommandsRegisterExt

Provides methods for managing the available “console commands”

CliCommandsRunExt

Provides methods to run/call “console commands”

Component

A data type that can be used to store data for an entity.

Condition

A system that determines if one or more scheduled systems should run.

Context

Provides the context method for Result.

CryptoRng

A marker trait used to indicate that an RngCore or BlockRngCore implementation is supposed to be cryptographically secure.

CubicGenerator

Implement this on cubic splines that can generate a curve from their spline parameters.

Debug

? formatting.

Decodable

A type implementing this trait can be converted to a rodio::Source type. It must be Send and Sync, and usually implements Asset so needs to be TypeUuid, in order to be registered. Types that implement this trait usually contain raw sound data that can be converted into an iterator of samples. This trait is implemented for AudioSource. Check the example audio/decodable for how to implement this trait on a custom type.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

DespawnRecursiveExt

Trait that holds functions for despawning recursively down the transform hierarchy

DetectChanges

Types that can read change detection information. This change detection is controlled by DetectChangesMut types such as ResMut.

DetectChangesMut

Types that implement reliable change detection.

Distribution

Types (distributions) that can be used to create a random instance of T.

DynamicAsset

Any type implementing this trait can be assigned to asset keys as part of a dynamic asset collection.

DynamicAssetCollection

This traits describes types that contain asset configurations and can register them in the DynamicAssets resource.

Event

A type that can be stored in an Events<E> resource You can conveniently access events using the EventReader and EventWriter system parameter.

FromPrimitive

A generic trait for converting a number to a value.

FromReflect

A trait that enables types to be dynamically constructed from reflected data.

FromWorld

Creates an instance of the type this trait is implemented for using data from the supplied World.

Geometry

Structs that implement this trait can be drawn as a shape. See the shapes module for some examples.

GetField

A convenience trait which combines fetching and downcasting of struct fields.

GetTupleStructField

A convenience trait which combines fetching and downcasting of tuple struct fields.

Hash

A hashable type.

HierarchyQueryExt

An extension trait for Query that adds hierarchy related methods.

IntoChainOptionalSystem

IntoChainResultSystem

IntoSystem

Conversion trait to turn something into a System.

IntoSystemConfigs

Types that can convert into a SystemConfigs.

IntoSystemSet

Types that can be converted into a SystemSet.

IntoSystemSetConfig

Types that can be converted into a SystemSetConfig.

IntoSystemSetConfigs

Types that can convert into a SystemSetConfigs.

IteratorRandom

Extension trait on iterators, providing random sampling methods.

LoadingStateAppExt

Extension trait for Bevy Apps to add loading states idiomatically

Material

Materials are used alongside MaterialPlugin and MaterialMeshBundle to spawn entities that are rendered with a specific Material type. They serve as an easy to use high level way to render Mesh entities with custom shader logic.

MaterialTilemap

Plugin

A collection of Bevy app logic and configuration.

PluginGroup

Combines multiple Plugins into a single unit.

ProgressSystem

Extension trait for systems with progress tracking

ReadOnlySystem

System types that do not modify the World when run. This is implemented for any systems whose parameters all implement ReadOnlySystemParam.

Reflect

The core trait of bevy_reflect, used for accessing and modifying data dynamically.

Resource

A type that can be inserted into a World as a singleton.

Rng

An automatically-implemented extension trait on RngCore providing high-level generic methods for sampling values and other convenience methods.

RngCore

The core of a random number generator.

SeedableRng

A random number generator that can be explicitly seeded.

Serialize

A data structure that can be serialized into any data format supported by Serde.

SliceRandom

Extension trait on slices, providing random mutation and sampling methods.

States

Types that can define world-wide states in a finite-state machine.

Struct

A trait used to power struct-like operations via reflection.

System

An ECS system that can be added to a Schedule

SystemParamFunction

A trait implemented for all functions that can be used as Systems.

SystemSet

Types that identify logical groups of systems.

ToPrimitive

A generic trait for converting a value to a number.

TransformPoint

A trait for point transformation methods.

TupleStruct

A trait used to power tuple struct-like operations via reflection.

Functions

any_component_removed

Generates a Condition-satisfying closure that returns true if there are any entity with a component of the given type removed.

any_with_component

Generates a Condition-satisfying closure that returns true if there are any entities with the given component type.

apply_deferred

Instructs the executor to call System::apply_deferred on the systems that have run but not applied their Deferred system parameters (like Commands) or other system buffers.

apply_state_transition

If a new state is queued in NextState<S>, this system:

chunk_aabb

Calculates the Aabb of a generic chunk. The AABB depends upon the grid size, tile size, and map type of the the chunk the tile belongs to.

chunk_index_to_world_space

Calculates the world-space position of the bottom-left of the specified chunk.

dbg

System adapter that utilizes the bevy_utils::tracing::debug! macro to print the output of a system.

default

An ergonomic abbreviation for Default::default() to make initializing structs easier. This is especially helpful when combined with “struct update syntax”.

despawn_all

Convenience system for despawning all entities that match a given query filter

despawn_all_recursive

despawn_with_children_recursive

Function for despawning an entity and all its children

error

System adapter that utilizes the bevy_utils::tracing::error! macro to print the output of a system.

fill_tilemap

Fills an entire tile storage with the given tile.

fill_tilemap_hexagon

Fills a hexagonal region with the given tile_texture.

fill_tilemap_rect

Fills a rectangular region with the given tile.

fill_tilemap_rect_color

Fills a rectangular region with colored versions of the given tile.

generate_hex_ring

Generates a vector of hex positions that form a ring of given radius around the specified origin.

generate_hexagon

Generates a vector of hex positions that form a hexagon of given radius around the specified origin.

get_tilemap_center_transform

Calculates a Transform for a tilemap that places it so that its center is at (0.0, 0.0, 0.0) in world space.

ignore

System adapter that ignores the output of the previous system in a pipe. This is useful for fallible systems that should simply return early in case of an Err/None.

in_state

Generates a Condition-satisfying closure that returns true if the state machine is currently in state.

info

System adapter that utilizes the bevy_utils::tracing::info! macro to print system information.

init_resource

Convenience system for initting a resource of the given type

not

Generates a Condition that inverses the result of passed one.

on_event

Generates a Condition-satisfying closure that returns true if there are any new events of the given type since it was last called.

random

Generates a random value using the thread-local random number generator.

remove_from_all

Convenience system for removing a component from all entities that match a given query filter

remove_resource

Convenience system for removing a resource of the given type

resource_added

Generates a Condition-satisfying closure that returns true if the resource of the given type has been added since the condition was last checked.

resource_changed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_changed_or_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_equals

Generates a Condition-satisfying closure that returns true if the resource is equal to value.

resource_exists

Generates a Condition-satisfying closure that returns true if the resource exists.

resource_exists_and_changed

Generates a Condition-satisfying closure that returns true if the resource of the given type has had its value changed since the condition was last checked.

resource_exists_and_equals

Generates a Condition-satisfying closure that returns true if the resource exists and is equal to value.

resource_removed

Generates a Condition-satisfying closure that returns true if the resource of the given type has been removed since the condition was last checked.

run_once

Generates a Condition-satisfying closure that returns true if the first time the condition is run and false every time after

state_changed

Generates a Condition-satisfying closure that returns true if the state machine changed state.

state_exists

Generates a Condition-satisfying closure that returns true if the state machine exists.

state_exists_and_equals

Generates a Condition-satisfying closure that returns true if the state machine exists and is currently in state.

thread_rng

Retrieve the lazily-initialized thread-local random number generator, seeded by the system. Intended to be used in method chaining style, e.g. thread_rng().gen::<i32>(), or cached locally, e.g. let mut rng = thread_rng();. Invoked by the Default trait, making ThreadRng::default() equivalent.

unwrap

System adapter that unwraps the Ok variant of a Result. This is useful for fallible systems that should panic in the case of an error.

warn

System adapter that utilizes the bevy_utils::tracing::warn! macro to print the output of a system.

Type Definitions

AnyResult

Result<T, Error>

AudioBundle

Bundle for playing a standard bevy audio asset

ColorMesh2dBundle

A component bundle for entities with a Mesh2dHandle and a ColorMaterial.

HashMap

A HashMap implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

HashSet

A HashSet implementing aHash, a high speed keyed hashing algorithm intended for use in in-memory hashmaps.

Instant

PbrBundle

A component bundle for PBR entities with a Mesh and a StandardMaterial.

RxShutdown

SpatialAudioBundle

Bundle for playing a standard bevy audio asset with a 3D position

TilemapBundle

TxShutdown

Attribute Macros

bevy_main

reflect_trait

A macro that automatically generates type data for traits, which their implementors can then register.

Derive Macros

AssetCollection

Derive macro for AssetCollection

Bundle

Component

Debug

Derive macro generating an impl of the trait Debug.

Deref

Implements Deref for structs. This is especially useful when utilizing the newtype pattern.

DerefMut

Implements DerefMut for structs. This is especially useful when utilizing the newtype pattern.

Deserialize

DynamicPlugin

Generates a dynamic plugin entry point function for the given Plugin type.

Error

Event

FromPrimitive

Derives num_traits::FromPrimitive for simple enums and newtypes.

FromReflect

Derives the FromReflect trait.

Hash

Derive macro generating an impl of the trait Hash.

Reflect

The main derive macro used by bevy_reflect for deriving its Reflect trait.

Resource

Serialize

States

SystemSet

Derive macro generating an impl of the trait SystemSet.

ToPrimitive

Derives num_traits::ToPrimitive for simple enums and newtypes.