Module minewars::prelude::helpers::filling

Functions

fill_tilemap

Fills an entire tile storage with the given tile.

fill_tilemap_hexagon

Fills a hexagonal region with the given tile_texture.

fill_tilemap_rect

Fills a rectangular region with the given tile.

fill_tilemap_rect_color

Fills a rectangular region with colored versions of the given tile.

generate_hex_ring

Generates a vector of hex positions that form a ring of given radius around the specified origin.

generate_hexagon

Generates a vector of hex positions that form a hexagon of given radius around the specified origin.