Module minewars::prelude::helpers::hex_grid::neighbors

Structs

HexNeighbors

Stores some data T associated with each neighboring hex cell, if present.

Enums

HexColDirection

Compass directions of a tile in hexagonal column-oriented coordinate systems (Column, ColumnEven, and ColumnOdd).

HexDirection

Neighbors of a hexagonal tile. Zero corresponds with East for row-oriented tiles, and North for column-oriented tiles. It might also correspond with a custom direction for hex maps which are oriented by a non-standard rotation.

HexRowDirection

Compass directions of a tile in hexagonal row-oriented coordinate systems (Row, RowEven, and RowOdd).

Constants

HEX_DIRECTIONS

Array of HexDirection variants.

HEX_OFFSETS

Offsets of tiles that lie in each HexDirection.