Type Alias minewars::prelude::tess::geometry_builder::SimpleBuffersBuilder

pub type SimpleBuffersBuilder<'l> = BuffersBuilder<'l, Point2D<f32, UnknownUnit>, u16, Positions>;

A BuffersBuilder that takes the actual vertex type as input.

Aliased Type

struct SimpleBuffersBuilder<'l> {
    buffers: &'l mut VertexBuffers<Point2D<f32, UnknownUnit>, u16>,
    first_vertex: u32,
    first_index: u32,
    vertex_offset: u32,
    vertex_constructor: Positions,
}

Fields
buffers: &'l mut VertexBuffers<Point2D<f32, UnknownUnit>, u16>first_vertex: u32first_index: u32vertex_offset: u32vertex_constructor: Positions
Implementations

impl<'l, OutputVertex, OutputIndex, Ctor> BuffersBuilder<'l, OutputVertex, OutputIndex, Ctor>where OutputVertex: 'l, OutputIndex: 'l,

pub fn new( buffers: &'l mut VertexBuffers<OutputVertex, OutputIndex>, ctor: Ctor ) -> BuffersBuilder<'l, OutputVertex, OutputIndex, Ctor>

pub fn with_vertex_offset( self, offset: u32 ) -> BuffersBuilder<'l, OutputVertex, OutputIndex, Ctor>

pub fn with_inverted_winding( self ) -> InvertWinding<BuffersBuilder<'l, OutputVertex, OutputIndex, Ctor>>

Consumes self and returns a builder with opposite triangle face winding.

pub fn buffers<'a, 'b>(&'b self) -> &'a VertexBuffers<OutputVertex, OutputIndex>where 'b: 'a,