Trait minewars::prelude::tess::path::traits::Build

pub trait Build {
    type PathType;

    // Required method
    fn build(self) -> Self::PathType;
}

Builds a path.

This trait is separate from PathBuilder and SvgPathBuilder to allow them to be used as trait object (which isn’t possible when a method returns an associated type).

Required Associated Types

type PathType

The type of object that is created by this builder.

Required Methods

fn build(self) -> Self::PathType

Builds a path object, consuming the builder.