Struct minewars::prelude::tokio::io::Stderr

pub struct Stderr {
    std: SplitByUtf8BoundaryIfWindows<Blocking<Stderr>>,
}

A handle to the standard error stream of a process.

Concurrent writes to stderr must be executed with care: Only individual writes to this AsyncWrite are guaranteed to be intact. In particular you should be aware that writes using write_all are not guaranteed to occur as a single write, so multiple threads writing data with write_all may result in interleaved output.

Created by the stderr function.

Examples

use tokio::io::{self, AsyncWriteExt};

#[tokio::main]
async fn main() -> io::Result<()> {
    let mut stderr = io::stdout();
    stderr.write_all(b"Print some error here.").await?;
    Ok(())
}

Fields
std: SplitByUtf8BoundaryIfWindows<Blocking<Stderr>>