Trait minewars::prelude::LoadingStateAppExt

pub trait LoadingStateAppExt {
    // Required methods
    fn add_loading_state<S>(
        &mut self,
        loading_state: LoadingState<S>
    ) -> &mut Self
       where S: States;
    fn add_collection_to_loading_state<S, A>(
        &mut self,
        loading_state: S
    ) -> &mut Self
       where S: States,
             A: AssetCollection;
    fn add_dynamic_collection_to_loading_state<S, C>(
        &mut self,
        loading_state: S,
        file: &str
    ) -> &mut Self
       where S: States,
             C: DynamicAssetCollection + Asset;
    fn init_resource_after_loading_state<S, A>(
        &mut self,
        loading_state: S
    ) -> &mut Self
       where S: States,
             A: Resource + FromWorld;
}

Extension trait for Bevy Apps to add loading states idiomatically

Required Methods

fn add_loading_state<S>(&mut self, loading_state: LoadingState<S>) -> &mut Selfwhere S: States,

Add a loading state to your app

fn add_collection_to_loading_state<S, A>( &mut self, loading_state: S ) -> &mut Selfwhere S: States, A: AssetCollection,

Add an AssetCollection to the LoadingState

The added collection will be loaded and inserted into your Bevy app as a resource.

    App::new()
        .add_loading_state(
          LoadingState::new(GameState::Loading)
            .continue_to_state(GameState::Menu)
        )
        .add_collection_to_loading_state::<_, AudioAssets>(GameState::Loading)
        .add_collection_to_loading_state::<_, ImageAssets>(GameState::Loading)

fn add_dynamic_collection_to_loading_state<S, C>( &mut self, loading_state: S, file: &str ) -> &mut Selfwhere S: States, C: DynamicAssetCollection + Asset,

Register files to be loaded as a certain type of DynamicAssetCollection

During the loading state, the given dynamic asset collections will be loaded and their content registered. This will happen before trying to resolve any dynamic assets as part of asset collections.

You need to register a loader for your asset type yourself. If you want to see some code, take a look at the custom_dynamic_assets example.

fn init_resource_after_loading_state<S, A>( &mut self, loading_state: S ) -> &mut Selfwhere S: States, A: Resource + FromWorld,

Add any FromWorld resource to be initialized after all asset collections are loaded.

    App::new()
        .add_loading_state(
          LoadingState::new(GameState::Loading)
            .continue_to_state(GameState::Menu)
        )
        .add_collection_to_loading_state::<_, TextureForAtlas>(GameState::Loading)
        .init_resource_after_loading_state::<_, TextureAtlasFromWorld>(GameState::Loading)