Trait minewars::prelude::SystemSet

pub trait SystemSet: DynHash + Debug + Send + Sync + 'static {
    // Required method
    fn dyn_clone(&self) -> Box<dyn SystemSet, Global>;

    // Provided methods
    fn system_type(&self) -> Option<TypeId> { ... }
    fn is_anonymous(&self) -> bool { ... }
}

Types that identify logical groups of systems.

Required Methods

fn dyn_clone(&self) -> Box<dyn SystemSet, Global>

Creates a boxed clone of the label corresponding to this system set.

Provided Methods

fn system_type(&self) -> Option<TypeId>

Returns Some if this system set is a SystemTypeSet.

fn is_anonymous(&self) -> bool

Returns true if this system set is an AnonymousSet.