Struct minewars::view::ViewTileData

pub struct ViewTileData {
    bytes: [u8; 2],
}

The per-tile data of a view.

This is a compact bitfield representation of the game state that needs to be cached/tracked for view switching in multi-view modes.

Since it is stored for all plids in the session, we don’t want to waste memory.

We don’t need all game state. Some can be discarded on view switch, or just kept live in ECS entities and hidden (like explosion effects, etc), some can be recomputed (vis levels, roads).

Fields
bytes: [u8; 2]
Implementations

impl ViewTileData

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl ViewTileData

pub const fn into_bytes(self) -> [u8; 2]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 2]) -> Self

Converts the given bytes directly into the bitfield struct.

impl ViewTileData

pub fn owner(&self) -> <B4 as Specifier>::InOut

Returns the value of owner.

pub fn owner_or_err( &self ) -> Result<<B4 as Specifier>::InOut, InvalidBitPattern<<B4 as Specifier>::Bytes>>

Returns the value of owner.

#Errors

If the returned value contains an invalid bit pattern for owner.

pub fn with_owner(self, new_val: <B4 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of owner set to the given value.

#Panics

If the given value is out of bounds for owner.

pub fn with_owner_checked( self, new_val: <B4 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of owner set to the given value.

#Errors

If the given value is out of bounds for owner.

pub fn set_owner(&mut self, new_val: <B4 as Specifier>::InOut)

Sets the value of owner to the given value.

#Panics

If the given value is out of bounds for owner.

pub fn set_owner_checked( &mut self, new_val: <B4 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of owner to the given value.

#Errors

If the given value is out of bounds for owner.

pub fn digit(&self) -> <B3 as Specifier>::InOut

Returns the value of digit.

pub fn digit_or_err( &self ) -> Result<<B3 as Specifier>::InOut, InvalidBitPattern<<B3 as Specifier>::Bytes>>

Returns the value of digit.

#Errors

If the returned value contains an invalid bit pattern for digit.

pub fn with_digit(self, new_val: <B3 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of digit set to the given value.

#Panics

If the given value is out of bounds for digit.

pub fn with_digit_checked( self, new_val: <B3 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of digit set to the given value.

#Errors

If the given value is out of bounds for digit.

pub fn set_digit(&mut self, new_val: <B3 as Specifier>::InOut)

Sets the value of digit to the given value.

#Panics

If the given value is out of bounds for digit.

pub fn set_digit_checked( &mut self, new_val: <B3 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of digit to the given value.

#Errors

If the given value is out of bounds for digit.

pub fn kind(&self) -> <TileKind as Specifier>::InOut

Returns the value of kind.

pub fn kind_or_err( &self ) -> Result<<TileKind as Specifier>::InOut, InvalidBitPattern<<TileKind as Specifier>::Bytes>>

Returns the value of kind.

#Errors

If the returned value contains an invalid bit pattern for kind.

pub fn with_kind(self, new_val: <TileKind as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of kind set to the given value.

#Panics

If the given value is out of bounds for kind.

pub fn with_kind_checked( self, new_val: <TileKind as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of kind set to the given value.

#Errors

If the given value is out of bounds for kind.

pub fn set_kind(&mut self, new_val: <TileKind as Specifier>::InOut)

Sets the value of kind to the given value.

#Panics

If the given value is out of bounds for kind.

pub fn set_kind_checked( &mut self, new_val: <TileKind as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of kind to the given value.

#Errors

If the given value is out of bounds for kind.

pub fn item(&self) -> <ItemKind as Specifier>::InOut

Returns the value of item.

pub fn item_or_err( &self ) -> Result<<ItemKind as Specifier>::InOut, InvalidBitPattern<<ItemKind as Specifier>::Bytes>>

Returns the value of item.

#Errors

If the returned value contains an invalid bit pattern for item.

pub fn with_item(self, new_val: <ItemKind as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of item set to the given value.

#Panics

If the given value is out of bounds for item.

pub fn with_item_checked( self, new_val: <ItemKind as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of item set to the given value.

#Errors

If the given value is out of bounds for item.

pub fn set_item(&mut self, new_val: <ItemKind as Specifier>::InOut)

Sets the value of item to the given value.

#Panics

If the given value is out of bounds for item.

pub fn set_item_checked( &mut self, new_val: <ItemKind as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of item to the given value.

#Errors

If the given value is out of bounds for item.

pub fn has_structure(&self) -> <bool as Specifier>::InOut

Returns the value of has_structure.

pub fn has_structure_or_err( &self ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of has_structure.

#Errors

If the returned value contains an invalid bit pattern for has_structure.

pub fn with_has_structure(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of has_structure set to the given value.

#Panics

If the given value is out of bounds for has_structure.

pub fn with_has_structure_checked( self, new_val: <bool as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of has_structure set to the given value.

#Errors

If the given value is out of bounds for has_structure.

pub fn set_has_structure(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of has_structure to the given value.

#Panics

If the given value is out of bounds for has_structure.

pub fn set_has_structure_checked( &mut self, new_val: <bool as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of has_structure to the given value.

#Errors

If the given value is out of bounds for has_structure.

pub fn structure(&self) -> <StructureKind as Specifier>::InOut

Returns the value of structure.

pub fn structure_or_err( &self ) -> Result<<StructureKind as Specifier>::InOut, InvalidBitPattern<<StructureKind as Specifier>::Bytes>>

Returns the value of structure.

#Errors

If the returned value contains an invalid bit pattern for structure.

pub fn with_structure( self, new_val: <StructureKind as Specifier>::InOut ) -> Self

Returns a copy of the bitfield with the value of structure set to the given value.

#Panics

If the given value is out of bounds for structure.

pub fn with_structure_checked( self, new_val: <StructureKind as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of structure set to the given value.

#Errors

If the given value is out of bounds for structure.

pub fn set_structure(&mut self, new_val: <StructureKind as Specifier>::InOut)

Sets the value of structure to the given value.

#Panics

If the given value is out of bounds for structure.

pub fn set_structure_checked( &mut self, new_val: <StructureKind as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of structure to the given value.

#Errors

If the given value is out of bounds for structure.